Fix #149249: Auto-validate meshes during Alembic import if we detect bad faces
This extends the current check that Alembic does with a more complete
one that happens to mirror what was done for USD earlier in the year.

The current check only looked at the prior vertex to see if it was
duplicated while the new check checks for duplicates across the entire
face.

Pull Request: https://projects.blender.org/blender/blender/pulls/149788