Fix: Vulkan: Incorrect framebuffer blit size
This was detected when implementing support for KosmicKrisp. That driver would
assert when a blit command would write to zero pixels. This is a spec violation
but only happens in that case due an incorrect swapchain size when drawing to
a retina display.

The incorrect size happens as VK_EXT_metal_surface expects a different order
of events than Blender uses in order to initialize a retina display.

This PR only ensures that the incorrect blit command isn't send to the rendergraph.

Pull Request: https://projects.blender.org/blender/blender/pulls/150155