Fix #136609: Crashes when texture painting non-manifold geometry
When using the experimental texture painting feature on a non-manifold
mesh with 3+ faces connected to an edge, Blender crashes.

There are a few issues that cause this. The first issue occurs when
initializing mesh data for texture painting. When setting up data for
edges and their connected faces, a Map collection is used to store edge
indexes and the vertices making up the edge. The problem is that the
index gets shifted when adding or modifying this data which causes
corruption on non-manifold meshes with edges that have 3+ faces. There
can be different edge indexes with the same data, meaning an edge or
more has the vertices to another edge. This causes a crash when UV
Island data is created since it goes through assertions to ensure that
faces associated with edges have the appropriate vertices.

The second issue comes from an edge case in the current system for
solving seam bleeding for texture painting. The current system uses
Fans to find how meshes connect in 3D space, which is then used to add
seams to UV border edges. In the system's current state, it only works
with manifold fans since there isn't any logic to add seams to
non-manifold Fans at the moment. The problem is that meshes with 3+
faces to an edge can be marked as manifold since there aren't any
explicit checks for this situation, and the system commonly skips over
the extra edges. This then causes crashes further on in the process of
fixing seam bleeding since these Fans fail assertions.

The fix for the first issue is simply removing logic to shift
edge indexes. This prevents data corruption for non-manifold meshes,
and from my testing, it doesn't cause any issues with regular meshes
or other types of non-manifold meshes either.

The fix for the second issue is to add a case during fan creation to
check each edge as the system iterates over edges of a fan. If an edge
in the fan isn't manifold, the fan is marked as non-manifold.

Another solution that was considered to fix the second issue, was
counting how many segments were added to a fan and then checking how
many faces connect to the vertex the fan was created from to check that
all fan segments were found. This however is a slower and messier
solution since it would require nested loops to find the faces connected
to a vertex, and it doesn't result in more accurate results than
simply checking if all edges in a fan are manifold.

Note: Issue #97352 has more information about the Fan system for seam
bleeding.

Pull Request: https://projects.blender.org/blender/blender/pulls/147244