Fix #146596: OpenGL Subprocess Compilation: `ERROR (17): SharedMemory : handle_: File exists`
Fix a race condition with pipe_id.

Add a time_id to the shared memory identifier to prevent duplicates in
case of a crash.
This assumes that the shared memory won't be unlinked properly if
Blender crashes, so a handle with the same name may still be open if a
new Blender session launches with the same PID.

Pull Request: https://projects.blender.org/blender/blender/pulls/150136