Sculpt: Change boundary neighbor detection to include boundary edges
When performing smoothing or other operations that look at mesh
neighbors, the expectation is that only other boundary elements are
included to avoid shrinking the boundary.

Currently, this is achieved by checking if the adjacent element is
itself a boundary element. This is insufficient in the case of a strip
of faces, as the edge spanning two boundary elements may not itself be
a boundary edge.

To fix this, when calculating boundary vertices, also populate a
runtime cache stored on the `SculptSession` as a `Set<OrderedEdge>`
where inclusion in this cache indicates that an edge is a boundary
edge.

This also affects multires elements which are made from a strip of
base faces, as the high resolution vertex boundaries are based on
the base mesh.

Then, this cache can be used when needed during brush strokes for
the smooth, relax, and boundary brushes as well as the expand tool
and automasking detection.

From a performance perspective, for typical manifold meshes this
has negligible memory and computational implications, as most
elements are not boundary elements and thus do not take up space
in the `OrderedEdge` set.

Finally, unit tests were added for related methods to cover common
usecases for both constructing the boundary information and using it
when calculating neighbors.

Addresses: #127769, #127734

Pull Request: https://projects.blender.org/blender/blender/pulls/150113