UI: Project Icon
This icon represents a series of post-its on a board.
They go from right-to-left because:

* I wanted to give the icon some horizontal element.
* It is inspired on User Journey Maps (my first contant
to actually using post-its on the wall ;)

This icon is a variation of the desktop icon.
It is intended now as a placeholder, but honestly it works
as the final icon if needs be.

This icon is not used at the moment. It will be first
used by the project setup branch (project_awareness).

This is one of the missing icons listed on #148207.

Pull Request: https://projects.blender.org/blender/blender/pulls/149901