GPU: Parse shader resource declaration
This allows to use a more pleasing syntax for resource declaration.

Using macros is simpler than parsing but it comes with some caveats:
- Not easy to debug
- Not easy to read
- Not easy to understand the logic
- Impossible to do some mutations (i.e. class member declaration
  and usage in a function)

This patch make it so that we can use a C++ valid syntax for declaring
the shader resources. This syntax should be valid for the C++ shader
compilation.

Using the parser we issue the create info syntax for compilation
of the GPU module. In the future we can also create reflection data
which would be much harder (or impossible) to do with simple macros.

Note: This patch only supports the new syntax but doesn't make use
of it. The port will be progressive.

Rel #137446

Follow up task:
- Pipelines and Shader Entry point
- Add support for resource_table as objects

Pull Request: https://projects.blender.org/blender/blender/pulls/145097