Anim: reserve size for evaluation result map
Since we know how big the `Map` has to be we
can reserve the size in the constructor. This
increases performance by not having to move memory
around as the map size increases.

The higher the FCurve count, the larger the improvement.

| - | main | PR |
| - | - | - |
| rain | 0.0107s | 0.0104s |
| high_bone_count_armature | 0.0164s | 0.0144s |

part of #149405

Pull Request: https://projects.blender.org/blender/blender/pulls/150127