Refactor: Simplify compacted edit mesh undo storage
Replace the linked list implementation of `BArrayCustomData` with a Map
for data stored with array-store and a map for layers stored with
implicit sharing. The main purpose is to reduce the direct binding of
this code with CustomData, as part of #122398, but the code also becomes
more locally readable too.

Pull Request: https://projects.blender.org/blender/blender/pulls/150522