Fix #150289: Linked actions being reused
When linking in an object that already had an action assigned,
keyframing its data would use that action and add a new slot to it.
However that data would be lost on save and reload since slots
cannot be added in a library override.

The fix is to not share the action in such cases

Pull Request: https://projects.blender.org/blender/blender/pulls/150580