Fix #144299: Bone Eyedropper missing for "Override Transform" property
This was an oversight in the original implementation of the bone eyedropper.
Using the bone eyedropper was not possible for certain properties that wanted
a pose bone. The issue was that the poll and  sampling functions were not set
up to handle objects of type pose bone.
That is handled now.

Original PR: #121523

Pull Request: https://projects.blender.org/blender/blender/pulls/148918