Vulkan: Dynamic State front face
Based on the (flipped) scaling of an object the front facing can be
different. This lead to additional pipelines when you duplicated and
flipped the object using scaling.

This PR adds support to make the front face a dynamic part of the
pipeline resulting in less pipeline variations. This will reduce the
number of pipeline variations needed in !150583

Front face is part of `VK_EXT_extended_dynamic_state`  which is core in
Vulkan 1.3 contexts. All our platforms have support for it, but just in
case we add it as optional.

Pull Request: https://projects.blender.org/blender/blender/pulls/150747