Fix #150582: Compositing applied to render when compositing is disabled in render pipeline
Also fixes the compositor overriding the render result when sequencer is
enabled.

The interactive compositor runs independently of the 'Compositing'
option in the postprocessing panel, which caused the bug.
To fix this, the Group Output node is diregarded in the compositing
root tree to avoid overwriting the final render result.

Pull Request: https://projects.blender.org/blender/blender/pulls/150721