Fix: Outliner: Moving collections deletes them unexpectedly
Happens if the selected/dragged collections included the collection we
want to drop before/after/into.

Instead of tweaking a bunch of logic (which honestly I havent followed
all through) regarding what should happen if we are dragging a bunch of
collections into one that is itself selected, I think the safest/sanest
way to treat such scenario is to **ignore** a collection (that is itself
selected) as a drop target.

Part of #150676

Pull Request: https://projects.blender.org/blender/blender/pulls/150771