Grease Pencil: Allow all remaining sculpt tools to sculpt Bezier handles
This makes it so that many of the sculpt tools properly deform the
handles of `Beizer` curves.

This PR implements the following tools:
`Randomize`, `Grab`, `Pinch`, `Pull`, `Twist` and `Clone`.

The `Strength` and `Thickness` tools already work, and the `Smooth`
is implemented in #141912 because it has different logic.

With both of these PR, all sculpt mode tools will work with
`Bezier` handles.

Pull Request: https://projects.blender.org/blender/blender/pulls/141928