Refactor: Vulkan: Vertex Input Description Pool
Vertex inputs consist out of bindings and attributes. For each draw call the
data is extracted from geometry and the active shader. This PR is a
refactoring to place the bindings and attributes in a pool so in the future we
can read them without passing around the full description.

This allows integration of `VK_EXT_dynamic_vertex_input` extensions where
the vertex inputs will not be provided as part of the graphics pipeline, but
inside the command buffer.

Pipeline lookup should be a bit faster as complex hashing has been replaced
by a simpler hash lookup.

Pull Request: https://projects.blender.org/blender/blender/pulls/150908