Fix #145063: Crash when using redo panel with transform tool
Internally, the transform tools use some portion of the Mesh Filter
code. When this data is initialized, it attempts to get the current
location of the cursor for some calculations. This process uses the
brush, which may be null. Dependent code did not correctly check for
the brush being present, leading to a segmentation fault.

Pull Request: https://projects.blender.org/blender/blender/pulls/150931