Fix #150705: Relative multires navigation doesn't work with unselected object
When using the `object.subdivision_set` operator, it was possible for
there to be an active but unselected object (e.g. in Sculpt mode),
leading to the operator failing to execute. This is rather confusing
behavior, so it has been changed to fallback to the active object if
no selected objects exist.

Pull Request: https://projects.blender.org/blender/blender/pulls/150758