VSE: add Meta Strip Volume
It is currently not possible to change the volume of meta strips. The
current workflow would be to go into the meta strip and change the
volume of all sound strips in there, this can get pretty tedious when
there are multiple sound strips in the meta strip.

This PR solves this problem by implementing a volume property for the
meta strip, the volume of all strips in the meta strip get multiplied
by this value.

Ref: https://blender.community/c/rightclickselect/8O2o