Fix #149497: UV select linked fails to expand with shared vertex UV's
Selecting linked UV's with select-sync enabled wasn't expanding from
selected UV's when the vertex was part of another selected UV island.

Disable this behavior unless "Sticky Vertex" is enabled where the
behavior makes sense as in that case selecting an island selects all
UV's that share the vertex including UV's on other islands.

This made sense when select-sync implied "Sticky Vertex" but that
is no longer the case.