Refactor: Geometry Nodes: make evaluation graph independent of original node tree
Currently, one can easily run into crashes when closures/fields reference node
data and the source node tree is freed (by the depsgraph). While this should
never happen currently, one can easily run into this when using #145849 with
fields and closures being stored on bundles in geometries.  This patch makes the
evaluation data independent of the node tree it is built from. So when the node
tree is freed, its evaluation data can still be used.

At its core, this works by making another copy of the node tree that is owned by
the evaluation data. The node tree as well as the evaluation data use
`std::shared_ptr` so that fields and closures can take ownership of them. While
this additional copy is not ideal, it seems like the most practical solution.
Other solutions would require significantly more refactoring because there are
lots of places that reference node tree data.

Fortunately, node-tree copying hasn't really been a bottleneck yet and is still
optimizable (e.g. with #149545). So overall, that does not seem too bad given
the benefit of allowing workflows like in #150852.

Unfortunately, this can only really be tested with #145849, so for now the tests
are just checking whether we break existing setups. With #145849 we can also add
new regression tests that specifically test what this patch implements.

While I tested this patch a fair amount, there is still a certain amount of risk
that I didn't find all places that need to be updated. It should hopefully be
fairly straight forward to fix other cases when they come up.

Pull Request: https://projects.blender.org/blender/blender/pulls/150915