Fix: Potential crashes when using assets deleted from disk
blender/blender!150999 fixed most such cases when drag & dropping
assets, but it can still happen that importing an asset fails. For
example, when the .blend file exists, but the asset was deleted or
renamed inside of it.

Even after this change, in such rare cases this will still trigger the
drop operator which will then either fail or show a search menu to pick
the ID to operate on. Ideally we wouldn't run the operator at all. But
for now, just avoid the crashes.

In general, importing assets should always be expected to fail.

Pull Request: https://projects.blender.org/blender/blender/pulls/151000