Nodes: Node Wrangler: Add "Start Frame" property to "Add Image Sequence" Operator
This patch adds `frame_start` property on the `node.nw_add_sequence()`
operator as the title implies. Once the property is set, you don't have to
constantly keep changing the start frame of every image sequence you
add for that session. The default value is ,1 so the operator works as before
for those who don't want to change it.

This is especially helpful when importing image sequences that start on
1 hour mark, which is common in video editing and VFX workflows.

Pull Request: https://projects.blender.org/blender/blender/pulls/150448