Fix #151157: Material preview crash with skin modifier
The issue is that the mesh has the active UV map string set, but
no UV map attribute with the corresponding name. Regardless
of whether or not the mesh should be in that situation (for the
skin modifier it seems reasonable), the drawing extraction code
shouldn't crash. So add a "orco" fallback to the mesh tangent
calculation for use in case on of the UV maps is empty.

Pull Request: https://projects.blender.org/blender/blender/pulls/151198