EEVEE: Add Glossy reflection evaluation to planar probe capture
This adds back a partial evaluation of the lighting for planar
light-probe capture.

The previous version was only evaluating every BSDF as diffuse or
translucent. This gave glass object a very dull and unrealistic look
in sharp reflections.

This commit only evaluate 2 lobes for reflections and 2 for
transmission with one of each always considered fully diffuse.

The (non diffuse) glossy and refraction BSDFs have their
parameter averaged. For instance; a coating layer will be
averaged with the main reflection lobe of a car paint surface.

This offers a good compromise in look, performance and shader
variant compilation time.

The alternative would be porting the whole evaluation from the
deferred pipeline would be very detrimental for performance and
code management perspective.

Pull Request: https://projects.blender.org/blender/blender/pulls/150781