Fix #150553: Cycles HIP RT crash without rocm-hip-devel on Fedora
Both HIP and HIP RT rely on the HIP runtime, which is provided by the rocm-hip
package as both a versioned and an unversioned library (libamdhip64.so.x and
libamdhip64.so). On Fedora, however, rocm-hip ships only the versioned library,
while the unversioned one is provided by rocm-hip-devel.

The version of HIP RT used in Blender currently looks only for the unversioned
library and fails to load the driver if it’s missing. HIP itself, by contrast,
looks for both the versioned and unversioned libraries. Cycles enables HIP RT
as long as the hiprt library is present, even if HIP RT fails to load the
runtime—leading to undefined behavior.

This addresses the issue by disabling HIP RT when the unversioned HIP runtime
library is not found.

Ref #150553

Pull Request: https://projects.blender.org/blender/blender/pulls/151081