Nodes: Additional check for availability of next socket before adding spacing
Spacing was being added if there was another output socket but it didn't
check if was available in the first place. For example: For a VectorMath node,
output socket `Value` is not available if the node is outputting a `Vector`.

This patch now does the appropriate check.

Videos in the PR.

Co-authored-by: IREXTIA <developer@irextia.studio>
Pull Request: https://projects.blender.org/blender/blender/pulls/135882