Fix #151204: Crash when using compositor viewer
Blender crashes when using the compositor viewer in the interactive
compositor. This is because the code that writes to the viewer images
was not thread safe and the draw thread reads invalid memory in that
case. To fix this, we delay lock releasing until the viewer is fully
written.

Pull Request: https://projects.blender.org/blender/blender/pulls/151302